requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bidict==0.23.1
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
dnspython==2.7.0
eventlet==0.39.1
Flask==2.0.1
Flask-Compress==1.17
Flask-SocketIO==5.0.1
gevent==24.11.1
gevent-websocket==0.10.1
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
MarkupSafe==3.0.2
migrate==0.3.8
numpy==2.2.4
oauthlib==3.2.2
openai==1.65.5
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
proto-plus==1.26.1
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pyparsing==3.2.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.2.0
python-socketio==5.3.0
pytz==2025.1
ratelimit==2.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==76.0.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
tenacity==9.0.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
validators==0.34.0
Werkzeug==2.0.3
wsproto==1.2.0
zope.event==5.0
zope.interface==7.2
zstandard==0.23.0
